Update and rename RockPaperScissord.js to RockPaperScissors.js
Making the computer choice and comparing the players were wrapped in functions, and some variable names were changed because the jQuery layer was added and I wanted to avoid confusion. Some of the return values also had to be changed to make it easier to change some classes using jQuery. The  jQuery portion makes more sense when you see the HTML.